Provide DNS address for Staging Tasks
In some cases it is desirable for a Task to depend on resolution of its
own DNS address before it reports itself as RUNNING.  It's currently
impossible because DNS addresses are only generated for RUNNING Tasks.
This leads to the undesirable scenario of having to lie to Mesos in
order to get a DNS address generated.